+++ rancher2/provider.go
@@ -91,6 +91,7 @@ func Provider() terraform.ResourceProvider {
 
 		ResourcesMap: map[string]*schema.Resource{
 			"rancher2_app":                           resourceRancher2App(),
+			"rancher2_app_v2":                        resourceRancher2AppV2(),
 			"rancher2_auth_config_activedirectory":   resourceRancher2AuthConfigActiveDirectory(),
 			"rancher2_auth_config_adfs":              resourceRancher2AuthConfigADFS(),
 			"rancher2_auth_config_azuread":           resourceRancher2AuthConfigAzureAD(),
@@ -102,6 +103,7 @@ func Provider() terraform.ResourceProvider {
 			"rancher2_auth_config_ping":              resourceRancher2AuthConfigPing(),
 			"rancher2_bootstrap":                     resourceRancher2Bootstrap(),
 			"rancher2_catalog":                       resourceRancher2Catalog(),
+			"rancher2_catalog_v2":                    resourceRancher2CatalogV2(),
 			"rancher2_certificate":                   resourceRancher2Certificate(),
 			"rancher2_cloud_credential":              resourceRancher2CloudCredential(),
 			"rancher2_cluster":                       resourceRancher2Cluster(),
@@ -137,6 +139,7 @@ func Provider() terraform.ResourceProvider {
 		DataSourcesMap: map[string]*schema.Resource{
 			"rancher2_app":                           dataSourceRancher2App(),
 			"rancher2_catalog":                       dataSourceRancher2Catalog(),
+			"rancher2_catalog_v2":                    dataSourceRancher2CatalogV2(),
 			"rancher2_certificate":                   dataSourceRancher2Certificate(),
 			"rancher2_cloud_credential":              dataSourceRancher2CloudCredential(),
 			"rancher2_cluster":                       dataSourceRancher2Cluster(),