+++ rancher2/resource_rancher2_auth_config_activedirectory.go
@@ -5,7 +5,8 @@ import (
 	"log"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	managementClient "github.com/rancher/types/client/management/v3"
+	norman "github.com/rancher/norman/types"
+	managementClient "github.com/rancher/rancher/pkg/client/generated/management/v3"
 )
 
 func resourceRancher2AuthConfigActiveDirectory() *schema.Resource {
@@ -45,13 +46,35 @@ func resourceRancher2AuthConfigActiveDirectoryCreate(d *schema.ResourceData, met
 		if err != nil {
 			return fmt.Errorf("[ERROR] Checking to enable Auth Config %s: %s", AuthConfigActiveDirectoryName, err)
 		}
-	}
-
-	// Updated auth config
-	newAuth := &managementClient.ActiveDirectoryConfig{}
-	err = meta.(*Config).UpdateAuthConfig(auth.Links["self"], authActiveDirectory, newAuth)
-	if err != nil {
-		return fmt.Errorf("[ERROR] Updating Auth Config %s: %s", AuthConfigActiveDirectoryName, err)
+		// Updated auth config
+		authResource := &norman.Resource{
+			ID:      auth.ID,
+			Type:    auth.Type,
+			Links:   auth.Links,
+			Actions: auth.Actions,
+		}
+		testInput := &managementClient.ActiveDirectoryTestAndApplyInput{
+			ActiveDirectoryConfig: authActiveDirectory,
+			Enabled:               authActiveDirectory.Enabled,
+			Password:              d.Get("test_password").(string),
+			Username:              d.Get("test_username").(string),
+		}
+		err = client.APIBaseClient.Action(managementClient.ActiveDirectoryConfigType, "testAndApply", authResource, testInput, nil)
+		if err != nil {
+			return err
+		}
+	} else {
+		if len(auth.Actions["disable"]) > 0 {
+			err = client.Post(auth.Actions["disable"], nil, nil)
+			if err != nil {
+				return fmt.Errorf("[ERROR] Disabling Auth Config %s: %s", AuthConfigActiveDirectoryName, err)
+			}
+		}
+		newAuth := &managementClient.ActiveDirectoryConfig{}
+		err = meta.(*Config).UpdateAuthConfig(auth.Links["self"], authActiveDirectory, newAuth)
+		if err != nil {
+			return fmt.Errorf("[ERROR] Updating Auth Config %s: %s", AuthConfigActiveDirectoryName, err)
+		}
 	}
 
 	return resourceRancher2AuthConfigActiveDirectoryRead(d, meta)