+++ rancher2/resource_rancher2_multi_cluster_app.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	managementClient "github.com/rancher/types/client/management/v3"
+	managementClient "github.com/rancher/rancher/pkg/client/generated/management/v3"
 )
 
 func resourceRancher2MultiClusterApp() *schema.Resource {
@@ -234,8 +234,26 @@ func resourceRancher2MultiClusterAppDelete(d *schema.ResourceData, meta interfac
 		return fmt.Errorf(
 			"[ERROR] waiting for multi cluster app (%s) to be removed: %s", id, waitErr)
 	}
-
 	d.SetId("")
+
+	for i := range multiClusterApp.Targets {
+		client, err := meta.(*Config).ProjectClient(multiClusterApp.Targets[i].ProjectID)
+		if err != nil {
+			continue
+		}
+		mappID := splitProjectIDPart(multiClusterApp.Targets[i].ProjectID) + ":" + multiClusterApp.Targets[i].AppID
+		stateConf = &resource.StateChangeConf{
+			Pending:    []string{"removing"},
+			Target:     []string{"removed"},
+			Refresh:    appStateRefreshFunc(client, mappID),
+			Timeout:    d.Timeout(schema.TimeoutDelete),
+			Delay:      1 * time.Second,
+			MinTimeout: 3 * time.Second,
+		}
+		stateConf.WaitForState()
+	}
+	time.Sleep(5 * time.Second)
+
 	return nil
 }