+++ rancher2/structure_cluster_rke_config_monitoring.go
@@ -1,7 +1,7 @@
 package rancher2
 
 import (
-	managementClient "github.com/rancher/types/client/management/v3"
+	managementClient "github.com/rancher/rancher/pkg/client/generated/management/v3"
 )
 
 // Flatteners
@@ -55,8 +55,13 @@ func expandClusterRKEConfigMonitoring(p []interface{}) (*managementClient.Monito
 	if v, ok := in["provider"].(string); ok && len(v) > 0 {
 		obj.Provider = v
 	}
+	if obj.Provider != clusterRKEConfigMonitoringProviderDisabled {
+		// Setting Replicas to 1 if monitoring enabled
+		value := int64(1)
+		obj.Replicas = &value
+	}
 
-	if v, ok := in["replicas"].(int); ok {
+	if v, ok := in["replicas"].(int); ok && v > 0 {
 		value := int64(v)
 		obj.Replicas = &value
 	}