+++ rancher2/structure_cluster_template.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	managementClient "github.com/rancher/types/client/management/v3"
+	managementClient "github.com/rancher/rancher/pkg/client/generated/management/v3"
 )
 
 // Sorters
@@ -376,7 +376,7 @@ func expandClusterTemplateRevisions(p []interface{}) (int, []managementClient.Cl
 			obj[i].Enabled = &v
 		}
 
-		if v, ok := in["name"].(string); ok && len(v) > 0 {
+		if v, ok := in["name"].(string); ok {
 			obj[i].Name = v
 			names[v]++
 			if names[v] > 1 {